Refactor: searchbar: fix remaining occurrences of stale usage of searchbar redux files
