+++ src/app/reducer.ts
@@ -1,4 +1,4 @@
-import searchReducer from 'components/SearchBar/searchSlice';
+import searchReducer from 'components/SearchBar/redux/searchSlice';
 import currencyReducer from 'components/Currency/redux/currenciesDataSlice';
 
 // Combine all slice reducers here. No need to use combineReducers
